chore(files):reorganize files and folders
