+++ src/utils/handleBanner.ts
